fixed mysql specific datatype in migration

+++ db/migrate/20101126113939_add_sales_users_associations.rb
@@ -1,8 +1,8 @@
 class AddSalesUsersAssociations < ActiveRecord::Migration
   def self.up
     add_column :sales, :user_id, :integer
-    add_column :sales, :longitude, :double
-    add_column :sales, :latitude, :double
+    add_column :sales, :longitude, :float
+    add_column :sales, :latitude, :float
   end
 
   def self.down
